rename project and package
