+++ pom.xml
@@ -5,14 +5,14 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>3.4.0</version>
+		<version>3.4.1</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>ac.prsu.csit.mca2024</groupId>
-	<artifactId>mca-2024-tic-tac-toe-api</artifactId>
+	<groupId>in.ac.prsu.mca</groupId>
+	<artifactId>mca2025-tictactoeapi</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>mca-2024-tic-tac-toe-api</name>
-	<description>tic tac toe player API</description>
+	<name>mca2025-tictactoeapi</name>
+	<description>API to play and practise tic tac toe with computer, developed by MCA batch 2024-25</description>
 	<url/>
 	<licenses>
 		<license/>
@@ -55,6 +55,18 @@
 
 	<build>
 		<plugins>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<configuration>
+					<annotationProcessorPaths>
+						<path>
+							<groupId>org.projectlombok</groupId>
+							<artifactId>lombok</artifactId>
+						</path>
+					</annotationProcessorPaths>
+				</configuration>
+			</plugin>
 			<plugin>
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
