+++ src/main/java/in/ac/prsu/mca/mca2025_tictactoeapi/Mca2025TictactoeapiApplication.java
@@ -1,0 +1,13 @@
+package in.ac.prsu.mca.mca2025_tictactoeapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Mca2025TictactoeapiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Mca2025TictactoeapiApplication.class, args);
+	}
+
+}
