+++ src/main/java/in/ac/prsu/mca/mca2025_tictactoeapi/constants/PlayResult.java
@@ -1,0 +1,5 @@
+package in.ac.prsu.mca.mca2025_tictactoeapi.constants;
+
+public enum PlayResult {
+	IN_PROGRESS, DRAW, HUMAN_WIN , AI_WIN ;
+}
