+++ src/main/java/in/ac/prsu/mca/mca2025_tictactoeapi/pojo/ErrorResponse.java
@@ -1,0 +1,12 @@
+package in.ac.prsu.mca.mca2025_tictactoeapi.pojo;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ErrorResponse {
+	private String errorCode ;
+	private String errorMessage ;
+	private String errorDescription ;
+}
