+++ src/main/java/in/ac/prsu/mca/mca2025_tictactoeapi/pojo/GamePlayRequest.java
@@ -1,0 +1,11 @@
+package in.ac.prsu.mca.mca2025_tictactoeapi.pojo;
+
+import lombok.Data;
+
+@Data
+public class GamePlayRequest {
+	  private boolean startedByHuman ;
+	  private int position ;
+	  private int[][] gameState;
+	
+}
