+++ src/main/java/in/ac/prsu/mca/mca2025_tictactoeapi/pojo/GameState.java
@@ -1,0 +1,5 @@
+package in.ac.prsu.mca.mca2025_tictactoeapi.pojo;
+
+public class GameState {
+
+}
