+++ src/main/java/in/ac/prsu/mca/mca2025_tictactoeapi/pojo/NewGameRequest.java
@@ -1,0 +1,9 @@
+package in.ac.prsu.mca.mca2025_tictactoeapi.pojo;
+
+import lombok.Data;
+
+@Data
+public class NewGameRequest {
+  private boolean startByHuman ;
+  private int position ;
+}
