+++ src/main/java/in/ac/prsu/mca/mca2025_tictactoeapi/service/dto/DiagonalSum.java
@@ -1,0 +1,22 @@
+package in.ac.prsu.mca.mca2025_tictactoeapi.service.dto;
+
+public class DiagonalSum {
+    private int sumOfD1;
+	private int sumOfD2;
+	
+	
+	public int getSumOfD1() {
+		return sumOfD1;
+	}
+	public void setSumOfD1(int sumOfD1) {
+		this.sumOfD1 = sumOfD1;
+	}
+	public int getSumOfD2() {
+		return sumOfD2;
+	}
+	public void setSumOfD2(int sumOfD2) {
+		this.sumOfD2 = sumOfD2;
+	}
+	
+	
+}
