--- src/main/java/tic_tac_toe/Mca2024TicTacToeApiApplication.java
@@ -1,13 +1,0 @@
-package tic_tac_toe;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Mca2024TicTacToeApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Mca2024TicTacToeApiApplication.class, args);
-	}
-
-}
