--- src/main/java/tic_tac_toe/constants/PlayResult.java
@@ -1,5 +1,0 @@
-package tic_tac_toe.constants;
-
-public enum PlayResult {
-	IN_PROGRESS, DRAW, HUMAN_WIN , AI_WIN ;
-}
