--- src/main/java/tic_tac_toe/pojo/ErrorResponse.java
@@ -1,12 +1,0 @@
-package tic_tac_toe.pojo;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ErrorResponse {
-	private String errorCode ;
-	private String errorMessage ;
-	private String errorDescription ;
-}
