--- src/main/java/tic_tac_toe/pojo/GamePlayRequest.java
@@ -1,11 +1,0 @@
-package tic_tac_toe.pojo;
-
-import lombok.Data;
-
-@Data
-public class GamePlayRequest {
-	  private boolean startedByHuman ;
-	  private int position ;
-	  private int[][] gameState;
-	
-}
