--- src/main/java/tic_tac_toe/pojo/GameState.java
@@ -1,5 +1,0 @@
-package tic_tac_toe.pojo;
-
-public class GameState {
-
-}
