--- src/main/java/tic_tac_toe/pojo/NewGameRequest.java
@@ -1,9 +1,0 @@
-package tic_tac_toe.pojo;
-
-import lombok.Data;
-
-@Data
-public class NewGameRequest {
-  private boolean startByHuman ;
-  private int position ;
-}
