--- src/main/java/tic_tac_toe/service/validate/Count.java
@@ -1,17 +1,0 @@
-package tic_tac_toe.service.validate;
-
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@ToString
-public class Count {
-
-	private int aiMarkCount = 0 ;
-	private int humanMarkCount = 0;
-	
-	public Count(int aiMarkCount, int humanMarkCount){
-		this.aiMarkCount = aiMarkCount ;
-		this.humanMarkCount = humanMarkCount ;
-	}
-}
