--- src/main/java/tic_tac_toe/service/validate/ValidationService.java
@@ -1,9 +1,0 @@
-package tic_tac_toe.service.validate;
-
-import tic_tac_toe.pojo.GamePlayRequest;
-import tic_tac_toe.pojo.NewGameRequest;
-
-public interface ValidationService {
-	void isValidPlayRequest(GamePlayRequest gamePlayRequest) ;
-	void isValidNewGameRequest(NewGameRequest newGameRequest);
-}
