+++ src/main/resources/application.properties
@@ -1,2 +1,2 @@
-spring.application.name=mca-2024-tic-tac-toe-api
+spring.application.name=mca2025-tictactoeapi
 server.port=8086
