+++ src/test/java/in/ac/prsu/mca/mca2025_tictactoeapi/Mca2025TictactoeapiApplicationTests.java
@@ -1,0 +1,13 @@
+package in.ac.prsu.mca.mca2025_tictactoeapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Mca2025TictactoeapiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
