--- src/test/java/tic_tac_toe/Mca2024TicTacToeApiApplicationTests.java
@@ -1,13 +1,0 @@
-package tic_tac_toe;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Mca2024TicTacToeApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
